🚚 Rename: 도메인과 관련없는 컴포넌트 폴더 이동
